[Refactor] Move HDFS utilities into HDFS filesystem module
